// app/packages/cli/src/core/generated/blueprint/v1/schema.ts
export enum C4Level {
  C4_LEVEL_UNSPECIFIED = 0,
  C4_LEVEL_CONTEXT = 1,
  C4_LEVEL_CONTAINER = 2,
  C4_LEVEL_COMPONENT = 3,
  C4_LEVEL_CODE = 4,
  UNRECOGNIZED = -1,
}
// ...
export enum NodeType {
  NODE_TYPE_UNSPECIFIED = 0,
  NODE_TYPE_PERSON = 1,
  NODE_TYPE_SOFTWARE_SYSTEM = 2,
  NODE_TYPE_WEB_APP = 3,
  NODE_TYPE_MOBILE_APP = 4,
  NODE_TYPE_SINGLE_PAGE_APP = 5,
  NODE_TYPE_MICROSERVICE = 6,
  NODE_TYPE_DATABASE = 7,
  NODE_TYPE_CACHE_STORE = 8,
  NODE_TYPE_EVENT_BROKER = 9,
  NODE_TYPE_SERVERLESS_APP = 10,
  NODE_TYPE_COMPONENT = 11,
  NODE_TYPE_CODE_MODULE = 12,
  NODE_TYPE_RELATIONAL_DATABASE = 13,
  NODE_TYPE_GRPC_SERVICE = 14,
  NODE_TYPE_SERVERLESS_FUNCTION = 15,
  NODE_TYPE_REST_API = 16,
  NODE_TYPE_GATEWAY_API = 17,
  NODE_TYPE_BACKGROUND_WORKER = 18,
  UNRECOGNIZED = -1,
}
// ...
export enum DependencyType {
  DEPENDENCY_TYPE_UNSPECIFIED = 0,
  DEPENDENCY_TYPE_DIRECT_CALL = 1,
  DEPENDENCY_TYPE_PUBLISH_SUBSCRIBE = 2,
  DEPENDENCY_TYPE_READ_WRITE = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface SystemNode {
  id: string;
  type: NodeType;
  name: string;
  external?: boolean | undefined;
  properties: { [key: string]: any } | undefined;
  isTest?: boolean | undefined;
  x?: number | undefined;
  y?: number | undefined;
  entityRef?: string | undefined;
}

export interface SystemDependency {
  from: string;
  to: string;
  type: DependencyType;
  description?: string | undefined;
}

export interface SystemSchema {
  name: string;
  version: string;
  level: C4Level;
  parentRef?: string | undefined;
  nodes: SystemNode[];
  dependencies: SystemDependency[];
}
// ...
export const SystemDependency: MessageFns<SystemDependency> = {
  encode(message: SystemDependency, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.from !== '') {
      writer.uint32(10).string(message.from);
    }
    if (message.to !== '') {
      writer.uint32(18).string(message.to);
    }
    if (message.type !== 0) {
      writer.uint32(24).int32(message.type);
    }
    if (message.description !== undefined) {
      writer.uint32(34).string(message.description);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SystemDependency {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSystemDependency();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.from = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.to = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.description = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SystemDependency {
    return {
      from: isSet(object.from) ? globalThis.String(object.from) : '',
      to: isSet(object.to) ? globalThis.String(object.to) : '',
      type: isSet(object.type) ? dependencyTypeFromJSON(object.type) : 0,
      description: isSet(object.description) ? globalThis.String(object.description) : undefined,
    };
  },

  toJSON(message: SystemDependency): unknown {
    const obj: any = {};
    if (message.from !== '') {
      obj.from = message.from;
    }
    if (message.to !== '') {
      obj.to = message.to;
    }
    if (message.type !== 0) {
      obj.type = dependencyTypeToJSON(message.type);
    }
    if (message.description !== undefined) {
      obj.description = message.description;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SystemDependency>, I>>(base?: I): SystemDependency {
    return SystemDependency.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SystemDependency>, I>>(object: I): SystemDependency {
    const message = createBaseSystemDependency();
    message.from = object.from ?? '';
    message.to = object.to ?? '';
    message.type = object.type ?? 0;
    message.description = object.description ?? undefined;
    return message;
  },
};

function createBaseSystemSchema(): SystemSchema {
  return { name: '', version: '', level: 0, parentRef: undefined, nodes: [], dependencies: [] };
}

export const SystemSchema: MessageFns<SystemSchema> = {
  encode(message: SystemSchema, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== '') {
      writer.uint32(10).string(message.name);
    }
    if (message.version !== '') {
      writer.uint32(18).string(message.version);
    }
    if (message.level !== 0) {
      writer.uint32(24).int32(message.level);
    }
    if (message.parentRef !== undefined) {
      writer.uint32(34).string(message.parentRef);
    }
    for (const v of message.nodes) {
      SystemNode.encode(v!, writer.uint32(42).fork()).join();
    }
    for (const v of message.dependencies) {
      SystemDependency.encode(v!, writer.uint32(50).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SystemSchema {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSystemSchema();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.version = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.level = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.parentRef = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.nodes.push(SystemNode.decode(reader, reader.uint32()));
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.dependencies.push(SystemDependency.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SystemSchema {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : '',
      version: isSet(object.version) ? globalThis.String(object.version) : '',
      level: isSet(object.level) ? c4LevelFromJSON(object.level) : 0,
      parentRef: isSet(object.parentRef)
        ? globalThis.String(object.parentRef)
        : isSet(object.parent_ref)
          ? globalThis.String(object.parent_ref)
          : undefined,
      nodes: globalThis.Array.isArray(object?.nodes)
        ? object.nodes.map((e: any) => SystemNode.fromJSON(e))
        : [],
      dependencies: globalThis.Array.isArray(object?.dependencies)
        ? object.dependencies.map((e: any) => SystemDependency.fromJSON(e))
        : [],
    };
  },

  toJSON(message: SystemSchema): unknown {
    const obj: any = {};
    if (message.name !== '') {
      obj.name = message.name;
    }
    if (message.version !== '') {
      obj.version = message.version;
    }
    if (message.level !== 0) {
      obj.level = c4LevelToJSON(message.level);
    }
    if (message.parentRef !== undefined) {
      obj.parentRef = message.parentRef;
    }
    if (message.nodes?.length) {
      obj.nodes = message.nodes.map(e => SystemNode.toJSON(e));
    }
    if (message.dependencies?.length) {
      obj.dependencies = message.dependencies.map(e => SystemDependency.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SystemSchema>, I>>(base?: I): SystemSchema {
    return SystemSchema.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SystemSchema>, I>>(object: I): SystemSchema {
    const message = createBaseSystemSchema();
    message.name = object.name ?? '';
    message.version = object.version ?? '';
    message.level = object.level ?? 0;
    message.parentRef = object.parentRef ?? undefined;
    message.nodes = object.nodes?.map(e => SystemNode.fromPartial(e)) || [];
    message.dependencies = object.dependencies?.map(e => SystemDependency.fromPartial(e)) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isObject(value: any): boolean {
  return typeof value === 'object' && value !== null;
}

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}